# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "RESTful API para gerenciamento de contas bancárias",
        "title": "Account Management API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Support",
            "url": "http://example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/health": {
            "get": {
                "description": "Retorna o status de saúde da API e suas dependências",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Verificar saúde da API",
                "responses": {
                    "200": {
                        "description": "API está operacional",
                        "schema": {
                            "$ref": "#/definitions/handler.HealthResponse"
                        }
                    },
                    "503": {
                        "description": "API indisponível",
                        "schema": {
                            "$ref": "#/definitions/handler.ServiceUnavailableError"
                        }
                    }
                }
            }
        },
        "/v1/accounts": {
            "post": {
                "description": "Cria uma nova conta bancária com CPF ou CNPJ",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Accounts"
                ],
                "summary": "Criar nova conta",
                "parameters": [
                    {
                        "description": "Dados da conta",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Conta criada com sucesso",
                        "schema": {
                            "$ref": "#/definitions/domain.Account"
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "schema": {
                            "$ref": "#/definitions/handler.BadRequestError"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/handler.InternalServerError"
                        }
                    }
                }
            }
        },
        "/v1/accounts/{accountId}": {
            "get": {
                "description": "Retorna os detalhes de uma conta específica",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Accounts"
                ],
                "summary": "Obter conta por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "ID da conta",
                        "name": "accountId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Conta encontrada",
                        "schema": {
                            "$ref": "#/definitions/domain.Account"
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "$ref": "#/definitions/handler.BadRequestError"
                        }
                    },
                    "404": {
                        "description": "Conta não encontrada",
                        "schema": {
                            "$ref": "#/definitions/handler.NotFoundError"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/handler.InternalServerError"
                        }
                    }
                }
            }
        },
        "/v1/transactions": {
            "post": {
                "description": "Cria uma nova transação bancária (débito/crédito)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Criar transação",
                "parameters": [
                    {
                        "description": "Dados da transação",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.createTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Transação criada com sucesso",
                        "schema": {
                            "$ref": "#/definitions/domain.Transaction"
                        }
                    },
                    "400": {
                        "description": "Erro de validação",
                        "schema": {
                            "$ref": "#/definitions/handler.BadRequestError"
                        }
                    },
                    "404": {
                        "description": "Conta não encontrada",
                        "schema": {
                            "$ref": "#/definitions/handler.NotFoundError"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/handler.InternalServerError"
                        }
                    }
                }
            }
        },
        "/v1/transactions/{transactionId}": {
            "get": {
                "description": "Retorna os detalhes de uma transação específica",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transactions"
                ],
                "summary": "Obter transação por ID",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "ID da transação",
                        "name": "transactionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transação encontrada",
                        "schema": {
                            "$ref": "#/definitions/domain.Transaction"
                        }
                    },
                    "400": {
                        "description": "ID inválido",
                        "schema": {
                            "$ref": "#/definitions/handler.BadRequestError"
                        }
                    },
                    "404": {
                        "description": "Transação não encontrada",
                        "schema": {
                            "$ref": "#/definitions/handler.NotFoundError"
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor",
                        "schema": {
                            "$ref": "#/definitions/handler.InternalServerError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Account": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "integer"
                },
                "document_number": {
                    "type": "string"
                }
            }
        },
        "domain.OperationType": {
            "type": "integer",
            "format": "int32",
            "enum": [
                1,
                2,
                3,
                4
            ],
            "x-enum-varnames": [
                "Purchase",
                "InstallmentPurchase",
                "Withdrawal",
                "Payment"
            ]
        },
        "domain.Transaction": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "integer"
                },
                "amount": {
                    "type": "number"
                },
                "operation_type_id": {
                    "$ref": "#/definitions/domain.OperationType"
                },
                "transaction_id": {
                    "type": "integer"
                }
            }
        },
        "handler.BadRequestError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "VALIDATION_ERROR"
                },
                "message": {
                    "type": "string",
                    "example": "document_number is required"
                }
            }
        },
        "handler.CreateAccountRequest": {
            "type": "object",
            "required": [
                "document_number"
            ],
            "properties": {
                "document_number": {
                    "type": "string",
                    "example": "12345678901"
                }
            }
        },
        "handler.HealthResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "database connection failed"
                },
                "status": {
                    "type": "string",
                    "example": "up"
                }
            }
        },
        "handler.InternalServerError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "INTERNAL_ERROR"
                },
                "message": {
                    "type": "string",
                    "example": "unexpected error on internal service"
                }
            }
        },
        "handler.NotFoundError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "NOT_FOUND"
                },
                "message": {
                    "type": "string",
                    "example": "account not found"
                }
            }
        },
        "handler.ServiceUnavailableError": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "database connection failed"
                },
                "status": {
                    "type": "string",
                    "example": "unavailable"
                }
            }
        },
        "handler.createTransactionRequest": {
            "type": "object",
            "required": [
                "account_id",
                "amount",
                "operation_type_id"
            ],
            "properties": {
                "account_id": {
                    "type": "integer",
                    "example": 123
                },
                "amount": {
                    "type": "number",
                    "example": 100.5
                },
                "operation_type_id": {
                    "type": "integer",
                    "example": 1
                }
            }
        }
    }
}